chore: LinkedListQueue를 Model Group으로 이동 #2
