+++ Calculator/Calculator/Model/LinkedListQueue.swift
@@ -1,0 +1,38 @@
+//
+//  LinkedListQueue.swift
+//  Calculator
+//
+//  Created by 이아리 on 2021/11/08.
+//
+
+import Foundation
+
+struct LinkedListQueue<T> {
+    private var list = LinkedList<T>()
+    
+    var count: Int {
+        return list.count
+    }
+    
+    var isEmpty: Bool {
+        return list.isEmpty
+    }
+    
+    var front: T? {
+        return list.first?.value
+    }
+    
+    func enqueue(_ element: T) {
+        list.append(element)
+    }
+    
+    @discardableResult
+    func dequeue() -> T? {
+        if list.isEmpty { return nil }
+        return list.remove(at: 0)
+    }
+    
+    func clear() {
+        list.removeAll()
+    }
+}
